--- a/Utilities/JIT.cpp
+++ b/Utilities/JIT.cpp
@@ -190,18 +190,34 @@ asmjit::JitRuntime& asmjit::get_global_runtime()
 	return g_rt;
 }
 
-asmjit::Label asmjit::build_transaction_enter(asmjit::X86Assembler& c, asmjit::Label fallback)
+void asmjit::build_transaction_enter(asmjit::X86Assembler& c, asmjit::Label fallback, const asmjit::X86Gp& ctr, uint less_than)
 {
 	Label fall = c.newLabel();
 	Label begin = c.newLabel();
 	c.jmp(begin);
 	c.bind(fall);
-	c.test(x86::eax, _XABORT_RETRY);
-	c.jz(fallback);
+
+	if (less_than < 65)
+	{
+		c.add(ctr, 1);
+		c.test(x86::eax, _XABORT_RETRY);
+		c.jz(fallback);
+	}
+	else
+	{
+		// Count an attempt without RETRY flag as 65 normal attempts and continue
+		c.not_(x86::eax);
+		c.and_(x86::eax, _XABORT_RETRY);
+		c.shl(x86::eax, 5);
+		c.add(x86::eax, 1); // eax = RETRY ? 1 : 65
+		c.add(ctr, x86::rax);
+	}
+
+	c.cmp(ctr, less_than);
+	c.jae(fallback);
 	c.align(kAlignCode, 16);
 	c.bind(begin);
 	c.xbegin(fall);
-	return begin;
 }
 
 void asmjit::build_transaction_abort(asmjit::X86Assembler& c, unsigned char code)