--- a/Utilities/JIT.h
+++ b/Utilities/JIT.h
@@ -43,7 +43,7 @@ namespace asmjit
 	asmjit::JitRuntime& get_global_runtime();
 
 	// Emit xbegin and adjacent loop, return label at xbegin
-	Label build_transaction_enter(X86Assembler& c, Label fallback);
+	void build_transaction_enter(X86Assembler& c, Label fallback, const X86Gp& ctr, uint less_than);
 
 	// Emit xabort
 	void build_transaction_abort(X86Assembler& c, unsigned char code);