--- a/rpcs3/Emu/Cell/Modules/sysPrxForUser.h
+++ b/rpcs3/Emu/Cell/Modules/sysPrxForUser.h
@@ -18,7 +18,7 @@ vm::ptr<void> _sys_memset(vm::ptr<void> dst, s32 value, u32 size);
 struct sys_lwmutex_t;
 struct sys_lwmutex_attribute_t;
 
-error_code sys_lwmutex_create(vm::ptr<sys_lwmutex_t> lwmutex, vm::ptr<sys_lwmutex_attribute_t> attr);
+error_code sys_lwmutex_create(ppu_thread& ppu, vm::ptr<sys_lwmutex_t> lwmutex, vm::ptr<sys_lwmutex_attribute_t> attr);
 error_code sys_lwmutex_lock(ppu_thread& CPU, vm::ptr<sys_lwmutex_t> lwmutex, u64 timeout);
 error_code sys_lwmutex_trylock(ppu_thread& CPU, vm::ptr<sys_lwmutex_t> lwmutex);
 error_code sys_lwmutex_unlock(ppu_thread& CPU, vm::ptr<sys_lwmutex_t> lwmutex);