--- a/rpcs3/Emu/Cell/Modules/sys_mempool.cpp
+++ b/rpcs3/Emu/Cell/Modules/sys_mempool.cpp
@@ -102,7 +102,7 @@ s32 sys_mempool_create(ppu_thread& ppu, vm::ptr<sys_mempool_t> mempool, vm::ptr<
 	condAttr->ipc_key = 0; // Also no idea what this is
 	strcpy_trunc(condAttr->name, "mp_c" + std::to_string(*mempool));
 
-	ret = sys_cond_create(condid, *mutexid, condAttr);
+	ret = sys_cond_create(ppu, condid, *mutexid, condAttr);
 	if (ret != CELL_OK)
 	{  // TODO: Better exception handling.
 		fmt::throw_exception("mempool %x failed to create condition variable", mempool);
@@ -126,7 +126,7 @@ void sys_mempool_destroy(ppu_thread& ppu, sys_mempool_t mempool)
 		idm::remove<memory_pool_t>(mempool);
 		sys_mutex_unlock(ppu, mutexid);
 		sys_mutex_destroy(mutexid);
-		sys_cond_destroy(condid);
+		sys_cond_destroy(ppu, condid);
 	}
 	else
 	{