--- a/rpcs3/Emu/Cell/Modules/sys_mempool.cpp
+++ b/rpcs3/Emu/Cell/Modules/sys_mempool.cpp
@@ -88,7 +88,7 @@ s32 sys_mempool_create(ppu_thread& ppu, vm::ptr<sys_mempool_t> mempool, vm::ptr<
 	attr->flags = 0; //  Also no idea what this is.
 	strcpy_trunc(attr->name, "mp_m" + std::to_string(*mempool));
 
-	error_code ret = sys_mutex_create(mutexid, attr);
+	error_code ret = sys_mutex_create(ppu, mutexid, attr);
 	if (ret != 0)
 	{ // TODO: Better exception handling.
 		fmt::throw_exception("mempool %x failed to create mutex", mempool);
@@ -125,7 +125,7 @@ void sys_mempool_destroy(ppu_thread& ppu, sys_mempool_t mempool)
 		sys_mutex_lock(ppu, memory_pool->mutexid, 0);
 		idm::remove<memory_pool_t>(mempool);
 		sys_mutex_unlock(ppu, mutexid);
-		sys_mutex_destroy(mutexid);
+		sys_mutex_destroy(ppu, mutexid);
 		sys_cond_destroy(ppu, condid);
 	}
 	else