… rpcs3/Emu/Cell/SPURecompiler.cpp
… rpcs3/Emu/Cell/SPUThread.cpp
… rpcs3/Emu/Cell/SPUThread.h
… rpcs3/Emu/Cell/lv2/sys_cond.cpp
… rpcs3/Emu/Cell/lv2/sys_cond.h
… rpcs3/Emu/Cell/lv2/sys_memory.cpp
… rpcs3/Emu/Cell/lv2/sys_mutex.cpp
… rpcs3/Emu/Cell/lv2/sys_mutex.h
… rpcs3/Emu/Cell/lv2/sys_net.cpp
… rpcs3/Emu/Memory/vm.cpp
… rpcs3/Emu/System.cpp
… rpcs3/Emu/System.h