Remove Accurate PUTLLC option.
Implement fallback path for SPU transactions.