@Whatcookie 2019-06-06
Persona 5 is now crashing with `{RSX [0x00f6548]} RSX: nv406e::semaphore_acquire has timed out. semaphore_address=0x403005A0`
[P5TSXPRFA.log](https://github.com/RPCS3/rpcs3/files/3263508/P5TSXPRFA.log)

Ninja gaiden sigma 2 hangs the same way as master does

Ninja gaiden sigma 1 also hangs with semaphore_acquire timeout
[NGS1TSXPRFA.log](https://github.com/RPCS3/rpcs3/files/3263518/NGS1TSXPRFA.log)