@Whatcookie 2019-06-06
P5 and sigma crashes seem fixed, but performance in P5 is still bad without accurate GETLLAR set. 

Sigma 2 still hangs in the same manner