@incognitoh 2019-06-07
Good news. Testing 0.0.6-9320fec4,  MotorStorm 2 (and others) are now working again with TSX + LLVM SPU Mega. For me, SPU LLVM Mega was broken a few weeks ago when TSX was enabled but this PR finally fixes it.

(i7 4790k Haswell)