@Whatcookie 2019-06-07
(with force_abort microcode on commit 92ec6ab)
This PR in P5: 7 fps
Master in P5: 30 fps
Master/this PR with accurate GETLLAR in P5: 42 fps


While running `perf stat -T` (quick way to write 1 to the force_abort msr)
Master: 0 fps until I cancel
This PR: 0.22 fps until I cancel

The fallback is working, it wasn't in earlier commits. It doesn't really seem like its very useful though...