@Sztefanol 2019-06-08
Journey and SotC both work now on LLVM on Haswell.