@legend800 2019-06-08
Yep, Motorstorm Apoc. is working again with LLVM + TSX. Thanks for the regression fix!