@Whatcookie 2019-06-12
As discussed outside of github, recent commits improve performance for latest microcode, and almost all games have better performance than tsx off again. However...

Demon souls seems to have an issue on this pr with TSX enabled: 
![flicker](https://user-images.githubusercontent.com/26352541/59324004-16b88b00-8cab-11e9-9dc9-959b04e9b1be.png)
geometry now flickers in and out in certain scenes. 
I tested 92ec6ab7c3d83061e6dda04ae4a8cae8a9b14306 as well (last commit before performance improvement) and its still occurring there. 

The flickering occurs every frame, and in a much more intense way after writing 1 to force_abort msr, so perhaps an issue with the fallback?
![flicker2](https://user-images.githubusercontent.com/26352541/59324305-5469e380-8cac-11e9-84c7-9865b602662e.png)