@psennermann 2019-06-17
Any news about this PR? On master build at the moment TSX+LLVM seems to be broken with every game...