@ruipin 2019-06-17
Demon's Souls is a weird case, since the game has a specific race condition (game code, not RPCS3 fault) which causes the flicker. The more inconsistent the frame pacing, the more easily it triggers. I suspect frame pacing with the TSX fallback path is more inconsistent than before.