@Whatcookie 2019-06-17
Isn't the flickering you're thinking of the full screen effect which appears with write colour buffers enabled? This seems different from that, since I used a patch which disables the dynamic exposure (removes the need for write colour buffers).