@legend800 2019-06-17
Yeah, a ton of games are broken with LLVM+TSX on Master (25% in my set), and both are needed for good perf in a bunch of them. This seems more like hotfix to me and to prevent false positives in other PRs. 

Hopefully the Demon Souls thing can be sorted out as I don't see any other regressions noted and this can get merged soon.