@Nekotekina 2019-06-17
Updated PR, please test for regressions.