@Whatcookie 2019-06-17
Demon's souls flickering is fixed, even when writing 1 to force_abort msr. I'll edit this post if I find any regressions.