@Nekotekina 2019-06-17
Oh yeas, Preferred SPU Threads should be set to Auto with this PR.