@psennermann 2019-06-17
For now, after some quick testing, Ninokuni, Persona 5 and Shadow of the Damned seem to be ok