@Dime333 2019-06-20
So does all games work best with SPU threads set to auto now or is the speed gain from using a specific number of SPU threads in some games lost with this build?