@legend800 2019-06-21
Ironically, I did just found a regression with this. NFL Blitz (maybe other ea titles?) suffered a massive speed drop:

0.0.6-8266
TSX on:  20 fps
TSX off: 50 fps

0.0.6-8262
TSX on:  60 fps

[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/3313218/RPCS3.log.gz)